+++ src/main/scala/empresita/positions/ExecutiveDirector.scala
@@ -3,7 +3,7 @@ import empresita.{Employee, Qualification}
 
 object ExecutiveDirector extends Director {
   override val hierarchy: Int = 6
-  override val salary: Float = 18000
+  override val salary: Double = 18000
 
   override def has_requirements(emp: Employee): Boolean = {
     true
