+++ src/main/scala/empresita/positions/OperationDirector.scala
@@ -3,7 +3,7 @@ import empresita.Employee
 
 object OperationDirector extends Director {
   override val hierarchy: Int = 5
-  override val salary: Float = 13000
+  override val salary: Double = 13000
 
   override def has_requirements(emp: Employee): Boolean = {
     return true
