+++ src/main/scala/empresita/positions/SectorDirector.scala
@@ -3,7 +3,7 @@ import empresita.Employee
 
 object SectorDirector extends Director {
   override val hierarchy: Int = 4
-  override val salary: Float = 9500
+  override val salary: Double = 9500
 
   override def has_requirements(emp: Employee): Boolean = {
     return true
